requirements.txt
numpy
opencv-python
tensorflow
scikit-learn
matplotlib
Pillow
streamlit